Use a sparse array for handling generated views
Because the schedule ids in the preferences-backed logic are always in a
range without gaps the old implementation where views are stored in a
list works since ids and list indices correspond automatically.
The schedule ids in the database-backed logic on the other hand will not
have that relationship to the list indices since there will be gaps
between assigned ids when entities are deleted in the database.
Therefore a map is more suitable for storing the views in the future.

This change exposes the problem where the `long` ids have to be cast to
`int` to fit into some of the AlarmManager methods later in the flow.
This will probably never be a problem in actual use cases (since you can
add and remove schedules a _lot_ of times before the long cannot be cast
to an int any longer) but should still be fixed.

+++ src/main/java/dk/jens/backup/schedules/db/Schedule.java
@@ -196,7 +196,7 @@ public class Schedule {
      * @return scheduling data object
      */
     public static Schedule fromPreferences(SharedPreferences preferences,
-                                           int number) throws SchedulingException {
+            long number) throws SchedulingException {
         final Schedule schedule = new Schedule();
         schedule.id = number;
         schedule.enabled = preferences.getBoolean(
